+++ acceptance/aws/gen_mocks.go
@@ -145,6 +145,10 @@ type lambdaMock struct {
 	basicMock
 }
 
+type mqMock struct {
+	basicMock
+}
+
 type rdsMock struct {
 	basicMock
 }