+++ aws/fetch/config.go
@@ -35,6 +35,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/kinesis"
 	"github.com/aws/aws-sdk-go-v2/service/kms"
 	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	"github.com/aws/aws-sdk-go-v2/service/mq"
 	"github.com/aws/aws-sdk-go-v2/service/rds"
 	"github.com/aws/aws-sdk-go-v2/service/redshift"
 	"github.com/aws/aws-sdk-go-v2/service/route53"
@@ -95,6 +96,7 @@ type AWSAPI struct {
 	Cognitoidentityprovider *cognitoidentityprovider.Client
 	Cognitoidentity         *cognitoidentity.Client
 	Kafka                   *kafka.Client
+	Mq                      *mq.Client
 	Elasticbeanstalk        *elasticbeanstalk.Client
 }