+++ template/internal/ast/gen_entities.go
@@ -30,6 +30,7 @@ var entities = map[Entity]struct{}{
 	"application":              {},
 	"appscalingpolicy":         {},
 	"appscalingtarget":         {},
+	"broker":                   {},
 	"bucket":                   {},
 	"buildproject":             {},
 	"cachecluster":             {},